Region-stamp da float a uint: TOLTO il limite ~7 corpi vivi [SHADER non verificato offline]
Il marchio anti-spuntone (region-stamp) era un FLOAT (mantissa 24 bit) → reggeva
solo ~7 corpi vivi insieme. Ora UINT (esatto fino a 2^32, bodyId×2^20 → ~4095 corpi):
- PlanetLodTree.RegionId() ritorna uint; nuovo array VisibleRegions (uint) + buffer
  _RegionOfInstance dedicato (confronto INTERO esatto, niente più margine 0.5 da float).
- _SlabRegion (compute+shader) da float a uint; per-nodo via SetInt, batch via asuint
  dei bit in misc.w (BitConverter.Int32BitsToSingle lato C#).
- SlabPool: guardia BodyId riportata da 7 a 4000 (paracadute di sanità, non un limite).

C# verificato col gate offline (pulito). Gli shader (PlanetSurfaceGPU/PlanetHeight*)
NON sono verificabili offline con Unity aperto → da confermare a gioco aperto.

+++ Assets/Resources/Shaders/PlanetHeightCore.hlsl
@@ -103,10 +103,11 @@ int   _NN;            // vertici per lato del nodo (nodeRes+1)
 int   _NSlabOff;      // offset (in vertici) del primo vertice della fetta nel pool condiviso
 // REGION-STAMP: il fill scrive l'id REGIONE nella fetta (insieme alla geometria) → il vertex shader smaschera una fetta
 // che tiene ancora la geometria di una regione vecchia (churn). _SlabRegion[indice fetta] = id; per-nodo da uniform,
-// batch da job.misc (z = indice fetta, w = id). float per la fetta = id esatto (≤ 2^23).
-RWStructuredBuffer<float> _SlabRegion;
-int   _SlabIndex;      // indice della fetta da marchiare (path per-nodo)
-float _SlabRegionId;   // id regione da scrivere (path per-nodo)
+// batch da job.misc (z = indice fetta, w = id reinterpretato nei bit del float → asuint). UINT = id esatto fino a 2^32
+// → niente più limite ~7 corpi vivi (era float, mantissa 24 bit). Confronto INTERO esatto nel vertex shader.
+RWStructuredBuffer<uint> _SlabRegion;
+int  _SlabIndex;      // indice della fetta da marchiare (path per-nodo)
+uint _SlabRegionId;   // id regione UINT da scrivere (path per-nodo; SetInt ne passa i bit)
 
 // ---- LOD BATCH (un SOLO dispatch per molti nodi): invece di settare gli uniform e fare 1 dispatch PER NODO
 // (~centinaia di chiamate API/frame nel churn), i parametri per-nodo stanno in un buffer e si dispatcha una
